Removing devices subpackage
