+++ hallbench/__init__.py
@@ -8,7 +8,6 @@ with open(_os.path.join(__path__[0], 'VERSION'), 'r') as _f:
 __all__ = (
     'calibration',
     'data',
-    'devices',
     'resources',
     'gui',
     )
