--- hallbench/devices/__init__.py
@@ -1,13 +1,0 @@
-"""Sub-package for devices communication."""
-
-from . import Agilent3458ALib
-from . import Agilent34401ALib
-from . import Agilent34970ALib
-from . import DRSLib
-from . import ElcomatLib
-from . import HeidenhainLib
-from . import NMRLib
-from . import PmacLib
-from . import UDCLib
-from . import interfaces
-from . import devices
